[dlocate] conver saveGob to methouds
