+++ dlocate/directoryPartition.go
@@ -1,6 +1,11 @@
 package main
 
-import "strings"
+import (
+	"os"
+	"strings"
+
+	log "github.com/Sirupsen/logrus"
+)
 
 // DirectoryPartition saves index of partitions
 type DirectoryPartition map[string]int
@@ -23,4 +28,28 @@ func (d DirectoryPartition) getDirectoryPartition(path string) int {
 		}
 	}
 	return -1
+}
+
+func readDirectoryPartitionGob() DirectoryPartition {
+	path := "indexFiles/directoryPartition.gob"
+
+	var directoryPartition DirectoryPartition
+	err := readGob(path, &directoryPartition)
+	if err != nil {
+		log.Error("Error while reading directoryPartition")
+		os.Exit(1)
+	}
+
+	return directoryPartition
+}
+
+func (d *DirectoryPartition) saveAsGob() {
+
+	path := "indexFiles/directoryPartition.gob"
+	err := saveGob(path, d)
+
+	if err != nil {
+		log.Errorf("Error while creating directoryPartitio file")
+		os.Exit(1)
+	}
 }
